+++ CHANGELOG.md
@@ -11,7 +11,9 @@
 - přesunutí obsahu z Introduction do druhé kapitoly
 - přidání krátké zmínky o Laplaceho aproximaci na konec kapitoly 2.3
 - přidání krátkého odstavce o výhodách bayes. metod díky hierarchickým modelům mezi kapitolami 2 a 2.1
-
+- přidána teoretická část
+    - TODO: Opravit z ARCH(m) na ARCH(p)
+    - TODO: Opravit na GARCH(p, q)
 ## Poznámky:
 
 - jinou citační formu použít nelze, musíme mít APA7 a je to v templatu